+++ examples/nlp/gpt/train_gpt_dpo.py
@@ -115,6 +115,7 @@ def main(cfg) -> None:
         load_gbs=True,
         pad_samples_to_global_batch_size=False,
         collate_fn=identity_collate,
+        limit_batches=cfg.trainer.dpo.limit_train_batches,
     )
 
     val_dataloader = build_dataloader(